requirements.txt
streamlit
pandas
pydeck
Pillow